agent/v1-sgoffice-proxmos-agent-a1/requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.20
pydantic>=2.4.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Proxmox integration
proxmoxer>=2.0.0
requests>=2.31.0

# Slack integration
slack-bolt>=1.18.0
slack-sdk>=3.24.0

# Google Calendar integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# LLM integrations
openai>=1.3.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0